feat(15) Simplified RootElements parsing
