+++ pkg/bpmn/model/bpmn20/process.go
@@ -2,50 +2,6 @@ package bpmn20
 
 import "github.com/pbinitiative/zenbpm/pkg/bpmn/model/extensions"
 
-type Activity interface {
-	FlowNode
-	GetCompletionQuantity() int
-	GetIsForCompensation() bool
-	GetStartQuantity() int
-}
-type TActivity struct {
-	TFlowNode
-	CompletionQuantity int  `xml:"completionQuantity,attr"`
-	IsForCompensation  bool `xml:"isForCompensation,attr"`
-	StartQuantity      int  `xml:"startQuantity,attr"`
-}
-
-func (act TActivity) GetCompletionQuantity() int {
-	return act.CompletionQuantity
-}
-func (act TActivity) GetIsForCompensation() bool {
-	return act.IsForCompensation
-}
-func (act TActivity) GetStartQuantity() int {
-	return act.StartQuantity
-}
-
-type TTask struct {
-	TActivity
-	// BPMN 2.0 Unorthodox elements. Part of the extensions elements
-	Input  []extensions.TIoMapping `xml:"extensionElements>ioMapping>input"`
-	Output []extensions.TIoMapping `xml:"extensionElements>ioMapping>output"`
-}
-
-// ExternallyProcessedTask is to be processed by external Job workers. Is not part of original BPMN Implementation
-// BPMN 2.0 Unorthodox.
-type TExternallyProcessedTask struct {
-	TTask
-	TaskDefinition extensions.TTaskDefinition `xml:"extensionElements>taskDefinition"`
-}
-
-type TEvent = TActivity
-
-type TGateway struct {
-	TFlowNode
-	GatewayDirection GatewayDirection `xml:"gatewayDirection,attr"`
-}
-
 type TFlowElementsContainer struct {
 	StartEvents            []TStartEvent             `xml:"startEvent"`
 	EndEvents              []TEndEvent               `xml:"endEvent"`
@@ -69,23 +25,125 @@ type TProcess struct {
 	DefinitionalCollaborationRef string `xml:"definitionalCollaborationRef,attr"`
 }
 
-func init() {
-	registeredRootElements[RootElementProcess] = func() RootElement { return &TProcess{} }
+type Activity interface {
+	FlowNode
+	GetCompletionQuantity() int
+	GetIsForCompensation() bool
+	GetStartQuantity() int
 }
 
-type FlowElementsType = []TFlowElement
-
-func (p *TProcess) getRootElementType() RootElementType { return RootElementProcess }
-
-func (d *RootElementsType) Process() *TProcess {
-	for _, re := range []RootElement(*d) {
-		if re.getRootElementType() == RootElementProcess {
-			return re.(*TProcess)
-		}
-	}
-	return nil
+type TActivity struct {
+	TFlowNode
+	CompletionQuantity int  `xml:"completionQuantity,attr"`
+	IsForCompensation  bool `xml:"isForCompensation,attr"`
+	StartQuantity      int  `xml:"startQuantity,attr"`
 }
+type TTask struct {
+	TActivity
+	// BPMN 2.0 Unorthodox elements. Part of the extensions elements
+	Input  []extensions.TIoMapping `xml:"extensionElements>ioMapping>input"`
+	Output []extensions.TIoMapping `xml:"extensionElements>ioMapping>output"`
+}
+
+type TGateway struct {
+	TFlowNode
+	GatewayDirection GatewayDirection `xml:"gatewayDirection,attr"`
+}
+
+type TEvent = TActivity
+
+// ExternallyProcessedTask is to be processed by external Job workers. Is not part of original BPMN Implementation
+// BPMN 2.0 Unorthodox.
+type TExternallyProcessedTask struct {
+	TTask
+	TaskDefinition extensions.TTaskDefinition `xml:"extensionElements>taskDefinition"`
+}
+type TServiceTask struct {
+	TExternallyProcessedTask
+	OperationRef   string `xml:"operationRef,attr"`
+	Implementation string `xml:"implementation,attr"`
+}
+
+// BPMN 2.0 Unorthodox elements. Part of the extensions elements
+type TUserTask struct {
+	TTask
+	AssignmentDefinition extensions.TAssignmentDefinition `xml:"extensionElements>assignmentDefinition"`
+}
+
+type TInclusiveGateway struct {
+	TGateway
+}
+
+type TParallelGateway struct {
+	TGateway
+}
+
+type TExclusiveGateway struct {
+	TGateway
+}
+
+type TEventBasedGateway struct {
+	TGateway
+	Instantiate      bool                  `xml:"instantiate,attr"`
+	EventGatewayType EventBasedGatewayType `xml:"eventGatewayType,attr"`
+}
+type EventBasedGatewayType = string
 
 const (
-	RootElementProcess RootElementType = "process"
+	EventBasedGatewayTypeExclusive EventBasedGatewayType = "Exclusive"
+	EventBasedGatewayTypeParallel  EventBasedGatewayType = "Parallel"
 )
+
+type TStartEvent struct {
+	TEvent
+	IsInterrupting   bool `xml:"isInterrupting,attr"`
+	ParallelMultiple bool `xml:"parallelMultiple,attr"`
+}
+
+type TEndEvent = TEvent
+
+type TIntermediateCatchEvent struct {
+	TEvent
+	MessageEventDefinition TMessageEventDefinition `xml:"messageEventDefinition"`
+	TimerEventDefinition   TTimerEventDefinition   `xml:"timerEventDefinition"`
+	LinkEventDefinition    TLinkEventDefinition    `xml:"linkEventDefinition"`
+	ParallelMultiple       bool                    `xml:"parallelMultiple"`
+	// BPMN 2.0 Unorthodox elements. Part of the extensions elements
+	Output []extensions.TIoMapping `xml:"extensionElements>ioMapping>output"`
+}
+
+type TIntermediateThrowEvent struct {
+	TEvent
+	LinkEventDefinition TLinkEventDefinition `xml:"linkEventDefinition"`
+	// BPMN 2.0 Unorthodox elements. Part of the extensions elements
+	Output []extensions.TIoMapping `xml:"extensionElements>ioMapping>output"`
+}
+
+type TMessageEventDefinition struct {
+	Id         string `xml:"id,attr"`
+	MessageRef string `xml:"messageRef,attr"`
+}
+
+type TTimerEventDefinition struct {
+	Id           string        `xml:"id,attr"`
+	TimeDuration TTimeDuration `xml:"timeDuration"`
+}
+
+type TLinkEventDefinition struct {
+	Id   string `xml:"id,attr"`
+	Name string `xml:"name,attr"`
+}
+
+type TTimeDuration struct {
+	XMLText string `xml:",innerxml"`
+}
+
+func (act TActivity) GetCompletionQuantity() int {
+	return act.CompletionQuantity
+}
+func (act TActivity) GetIsForCompensation() bool {
+	return act.IsForCompensation
+}
+func (act TActivity) GetStartQuantity() int {
+	return act.StartQuantity
+}
